+++ src/femic/pipeline/__init__.py
@@ -23,6 +23,7 @@ from femic.pipeline.pre_vdyp import (
     serialize_vdyp_prep_payload,
 )
 from femic.pipeline.stages import StageResult, run_legacy_subprocess
+from femic.pipeline.vdyp_sampling import nsamples_from_curves
 from femic.pipeline.vdyp_io import import_vdyp_tables, write_vdyp_infiles_plylyr
 from femic.pipeline.vdyp import build_vdyp_log_paths
 
@@ -41,6 +42,7 @@ __all__ = [
     "serialize_vdyp_prep_payload",
     "import_vdyp_tables",
     "write_vdyp_infiles_plylyr",
+    "nsamples_from_curves",
     "write_manifest",
     "build_vdyp_log_paths",
     "build_pipeline_run_config",